requirements.txt
zope.component==4.2.2
zope.event==4.1.0
zope.interface==4.1.3
python-consul==0.4.7
tornado==4.3
funcsigs==1.0.0
mock==1.3.0
